Add several event stream metadata-related options
These are all based on initial design feedback from the dashboard
team.

First, we have `--event-stream-application` and
`--event-stream-environment` flags. These are identifiers that are
required when running in event stream mode, and are embedded in every
event sent out from the Supervisor. They will help to organize data on
the dashboard.

In the current design, these apply to the Supervisor as a whole, and
not to specific services that may be running on that Supervisor. This
restriction may be loosened in future iterations.

There is some unfortunate conceptual overlap between the
`--application` and `--environment` flags that currently exist, and
are incorporated into service group names. *That* particular feature
hasn't been completely fleshed out, and may go away or change. The
admittedly verbose naming of these new options is meant to help
distinguish the two.

Next, we have the `--event-meta` option. This allows users to supply
arbitrary key-value metadata pairs which, like the environment and
application flags above, will be added to each event the Supervisor
sends out. This option can be supplied multiple times (once per
key-value pair), and all pairs will be collected into a single map.

Unlike the environment and application flags, these more generic
metadata are completely user-defined, and do not have special meaning
in the dashboard UI, apart from being available as filters, etc.

An example invocation:

```sh
HAB_FEAT_EVENT_STREAM=1 \
hab sup run \
  --event-stream-application=MY_APP \
  --event-stream-environment=PRODUCTION \
  --event-meta="department=finance" \
  --event-meta="division=payroll"
```

All of these options are only available when the event stream feature
is enabled. The application and environment options are required.

Signed-off-by: Christopher Maier <cmaier@chef.io>